rename sample to example
