+++ .gitignore
@@ -11,6 +11,6 @@
 .externalNativeBuild
 identsdk/src/main/java/com/twigbit/identsdk/util/Secrets\.kt
 
-sample/src/main/java/com/twigbit/identsdk/sample/Secrets\.kt
+example/src/main/java/com/twigbit/identsdk/example/Secrets\.kt
 
 \.idea/markdown-exported-files\.xml
