+++ example/build.gradle
@@ -15,7 +15,7 @@ android {
     }
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.twigbit.identsdk.sample"
+        applicationId "com.twigbit.identsdk.example"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode 1
