+++ example/src/androidTest/java/com/twigbit/identsdk/ExampleInstrumentedTest.kt
