+++ example/src/main/ic_launcher-web.png
