+++ example/src/main/java/com/twigbit/identsdk/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.twigbit.identsdk.sample
+package com.twigbit.identsdk.example
 
 import android.os.Bundle
 import com.twigbit.identsdk.dropinui.DropInRequest
