+++ example/src/main/res/layout/activity_independent_identification.xml
