+++ example/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
         android:fitsSystemWindows="true"
         tools:context=".MainActivity">
     <TextView android:layout_width="wrap_content" android:layout_height="wrap_content"
-              android:text="Ident SDK Sample"
+              android:text="Ident SDK Example"
               android:textSize="22sp"
               android:gravity="center"
               android:fontFamily="@font/nexa_bold"
