+++ example/src/main/res/mipmap-xhdpi/ic_launcher_foreground.png
