+++ example/src/main/res/mipmap-xxxhdpi/ic_launcher_foreground.png
