+++ example/src/main/res/values/colors.xml
