+++ example/src/main/res/values/ic_launcher_background.xml
